breakout geometry in its own package
