+++ app/src/main/java/com/scepticalphysiologist/dmaple/ui/camera/MappingRoiOverlay.kt
@@ -14,9 +14,9 @@ import android.view.View
 import android.view.WindowManager
 import androidx.lifecycle.MutableLiveData
 import com.scepticalphysiologist.dmaple.R
-import com.scepticalphysiologist.dmaple.etc.Edge
-import com.scepticalphysiologist.dmaple.etc.Frame
-import com.scepticalphysiologist.dmaple.etc.Point
+import com.scepticalphysiologist.dmaple.geom.Edge
+import com.scepticalphysiologist.dmaple.geom.Frame
+import com.scepticalphysiologist.dmaple.geom.Point
 import com.scepticalphysiologist.dmaple.etc.msg.MultipleChoice
 import com.scepticalphysiologist.dmaple.map.field.FieldImage
 import com.scepticalphysiologist.dmaple.map.field.FieldRoi
