+++ app/src/test/java/com/scepticalphysiologist/dmaple/ScrapBookTest.kt
@@ -1,8 +1,8 @@
 package com.scepticalphysiologist.dmaple
 
 
-import com.scepticalphysiologist.dmaple.etc.Frame
-import com.scepticalphysiologist.dmaple.etc.Point
+import com.scepticalphysiologist.dmaple.geom.Frame
+import com.scepticalphysiologist.dmaple.geom.Point
 
 import org.junit.jupiter.api.Test
 import org.robolectric.RobolectricTestRunner
